requirements.txt
streamlit==1.32.0
langchain==0.1.6
langchain-google-genai==0.0.4
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.2